k8s/vector-config-reloader/requirements.txt
kubernetes==34.1.0
pyyaml==6.0.3
